A working Cooley-Tukey FFT implemented
The error in the test in the last push has been fixed
by using a reference of an Image object in the tests functions
instead of a copy of the object.

The copy constructor has been made private in the Image class, so
that copying an Image object is prohibited.

The algorithms namespace contains a few functions that could
go into a separate "utility" namespace: initNewComplexArray,
isAPowerOfTwo and splitArrayToEvenAndOdd.

The algorithms namespace
could only contain the FFT in 1D and 2D and the inverse FFT and
later the Filtered Back Projection.

Tests are also up to date.